requirements.txt
google-genai
google-api-python-client
google-generativeai
python-dotenv
newspaper3k
tqdm
nltk
lxml_html_clean
scikit-learn 
matplotlib 
seaborn